BACKEND_POC/requirements.txt
mysqlclient==2.0.1
Django==3.1.1
django-silk==4.1.0
django-extensions==3.0.9
djangorestframework==3.11.1
drf-yasg==1.17.1
celery==4.2.0
kombu==4.6.6